better memo for unstable ownProps
